requirements.txt
requests==2.24.0
deepdiff==6.2.3
pyjwt==2.6.0
python-dotenv==0.21.0
dictdiffer==0.9.0
